% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EfficiencyGap.R
\name{EfficiencyGap}
\alias{EfficiencyGap}
\title{A Gerrymandering Function}
\usage{
EfficiencyGap(dataframe)
}
\arguments{
\item{dataframe}{This should be a dataframe containing the district number, republican
votes, democrat votes and independent votes, in that order.}
}
\description{
This function is for the determining the voting efficiency in a particular state.
returns a single percentage for the state. Anything greater than +/- 8 indicates 
gerrymandering in a state. Negative for republican, postive for democrat.
}
\examples{
EfficiencyGap(OhioDistricts)

dataframe ex.
district  R_votes D_votes I_votes
1          2000    1500     0
2          1500    2500     200
}
\keyword{gerrymandering}
