% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EfficiencyScrapping.R
\name{EfficiencyScrapping}
\alias{EfficiencyScrapping}
\title{A Gerrymandering Data Collection Function}
\usage{
EfficiencyScrapping()
}
\description{
This function is to allow the user to scrape the data from a website. This
provides all the raw data for determining the extent of gerrymandering during
the 2016 election.
}
\examples{
EfficiencyScrapping()

}
\keyword{gerrymandering,}
\keyword{scrapping}
\keyword{web}
