% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EfficiencyShell.R
\name{EfficiencyShell}
\alias{EfficiencyShell}
\title{A Gerrymandering Function}
\usage{
EfficiencyShell()
}
\description{
This function is to provide an interface into using this package. instead of calling 
each script independently it allows for the user to select option on how they would 
like to use the information.
}
\examples{
EfficiencyShell()
}
\keyword{gerrymandering}
